Modify: set background image
