+++ src/components/CardListComposer/style.ts
@@ -6,5 +6,5 @@ export const Container = styled.div`
   min-width: 272px;
   background-color: ${({ theme }) => theme.color.cardListGray};
   border-radius: ${({ theme }) => theme.borderRadius.card};
-  height: 100%;
+  height: 90px;
 `;
