+++ src/pages/BoardPage/index.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { useSelector } from "react-redux";
 import { useSearchParams } from "react-router-dom";
 import loadable from "@loadable/component";
 import { SEARCH_PARAMS_KEY } from "@/constants";
@@ -19,6 +20,7 @@ interface BoardProps {
 }
 
 const BoardPage: React.FC = () => {
+  const background = useSelector((state: { background: string }) => state.background);
   const [searchParams, setSearchParams] = useSearchParams(window.location.search);
   const searchKeyword = searchParams.get(SEARCH_PARAMS_KEY.SEARCH) ?? "";
 
@@ -28,7 +30,7 @@ const BoardPage: React.FC = () => {
   };
 
   return (
-    <S.Container>
+    <S.Container background={background}>
       <Header />
       <S.ContentLayout>
         <Sider />
