+++ src/pages/BoardPage/style.ts
@@ -1,16 +1,22 @@
 import { Layout as AntdLayout } from "antd";
 import styled from "styled-components";
 
-export const Container = styled(AntdLayout).attrs(() => ({
-  style: {
-    minHeight: "100vh",
-    backgroundColor: "transparent",
-  },
-}))``;
+interface ContainerProps {
+  background: string;
+}
+
+export const Container = styled.div<ContainerProps>`
+  height: 100vh;
+  background-image: url(${({ background }) => background});
+  background-size: cover;
+  background-position: center center;
+  background-repeat: no-repeat;
+`;
 
 export const ContentLayout = styled(AntdLayout).attrs(() => ({
   style: {
     backgroundColor: "transparent",
+    height: "94.4%",
   },
 }))``;
